wx2.9x64 was tripping over some format calls with "%1$d" and size_t. Since all string items are 'd', change the format to cast long/size_t to int. In addition, entry-validation was triggering when selecting a page in wxNotebook - use ChangeSelection instead of SetSelection

+++ AgilityBook/src/Win/WizardExcel.cpp
@@ -551,7 +551,9 @@ bool CWizardExcelImport::GetData(
 			wxFileName filename(m_FileName);
 			wxString msg = filename.GetFullPath();
 			ioProgress->SetCaption(msg);
-			wxString str = wxString::Format(_("IDS_READING_ROWSCOLS"), nRows, nCols);
+			wxString str = wxString::Format(_("IDS_READING_ROWSCOLS"),
+				static_cast<int>(nRows),
+				static_cast<int>(nCols));
 			ioProgress->SetMessage(str);
 			ioProgress->SetRange(1, nRows);
 			ioProgress->ShowProgress();
